Changed entity id from int to qint64
svn path=/trunk/KDE/kdepimlibs/; revision=788790

+++ imapparser_p.h
@@ -111,7 +111,7 @@ class AKONADIPROTOCOLINTERNALS_EXPORT ImapParser
       @param ok Set to false if the parsing failed.
       @param start Start parsing at this index.
     */
-    static int parseNumber( const QByteArray &data, int &result, bool *ok = 0, int start = 0 );
+    static int parseNumber( const QByteArray &data, qint64 &result, bool *ok = 0, int start = 0 );
 
     /**
       Quotes the given QByteArray.
